Remove ComplianceSystemsTable component
